Renaming folder ViperFpga to Fpga for simplicity
