+++ src/ArduinoViperFpga.h
@@ -10,7 +10,7 @@
  * INCLUDE
  **************************************************************************************/
 
-#include "ViperFpga/FpgaIo.h"
+#include "Fpga/FpgaIo.h"
 
 /**************************************************************************************
  * CLASS DECLARATION
